13.11.2022 Mandate ready & Discretion half ready

+++ WindowsEmulator/Desktop.xaml.cs
@@ -75,6 +75,13 @@ namespace WindowsEmulator
             journalAdmin.MouseDoubleClick += new MouseButtonEventHandler(OpenAdminPanel);
             FolderArea.Children.Add(journalAdmin);
             Canvas.SetLeft(journalAdmin, 125);
+            // Дискреционная модель
+            DisplayTile discretionAdmin = new DisplayTile();
+            discretionAdmin.Text = "Дискреционная панель";
+            discretionAdmin.Image = new BitmapImage(new Uri(@"Assets/AdminPanel.png", UriKind.Relative));
+            discretionAdmin.MouseDoubleClick += new MouseButtonEventHandler(OpenDiscretionPanel);
+            FolderArea.Children.Add(discretionAdmin);
+            Canvas.SetLeft(discretionAdmin, 250);
             // Папки пользователей
             try
             {
@@ -84,6 +91,8 @@ namespace WindowsEmulator
                     foreach (string folder in FolderList)
                     {
                         string[] FolderData = folder.Split('|');
+                        string[] access = folder.Split('_');
+                        string names = "";
                         DisplayTile newFolder = new DisplayTile();
                         newFolder.ContextMenu = new ContextMenu();
                         MenuItem item = new MenuItem();
@@ -95,13 +104,20 @@ namespace WindowsEmulator
                         item1.Click += new RoutedEventHandler(EditFolder);
                         newFolder.ContextMenu.Items.Add(item1);
                         newFolder.Text = FolderData[0];
-                        newFolder.Name = FolderData[1];
-                        newFolder.Tag = FolderData[2];
+                        newFolder.Tag = FolderData[1];
+
+                        for (int i = 1; i != access.Length; i++)
+                        {
+                            names += access[i] + "_";
+                        }
+                        names = names.Remove(names.Length - 1, 1);
+                        newFolder.Name = names;
+
                         newFolder.Image = new BitmapImage(new Uri(@"Assets/Folder.png", UriKind.Relative));
                         newFolder.MouseDoubleClick += new MouseButtonEventHandler(OpenFolder);
                         FolderArea.Children.Add(newFolder);
-                        Canvas.SetLeft(newFolder, (Convert.ToDouble(FolderData[3])));
-                        Canvas.SetTop(newFolder, (Convert.ToDouble(FolderData[4])));
+                        Canvas.SetLeft(newFolder, (Convert.ToDouble(FolderData[2])));
+                        Canvas.SetTop(newFolder, (Convert.ToDouble(FolderData[3])));
                     }
                 }
             }
@@ -115,66 +131,66 @@ namespace WindowsEmulator
         private void DeleteFolder(object sender, RoutedEventArgs e)
         {
             var obj = ((sender as MenuItem).Parent as ContextMenu).PlacementTarget as DisplayTile;
-            if ((_currentUser._level >= Convert.ToInt32(obj.Tag)) || _currentUser._username == "Admin")
-            {
-                FolderArea.Children.Remove(obj);
-                Directory.Delete(@"UserFolder\" + obj.Text);
-                string[] FoldersList = File.ReadAllLines(@"Folders.txt");
-                File.Delete(@"Folders.txt");
-                foreach (string folder in FoldersList)
+                if ((_currentUser._level >= Convert.ToInt32(obj.Tag)) || _currentUser._username == "Admin")
                 {
-                    string[] folderData = folder.Split('|');
-                    if (folderData[0] != obj.Text)
+                    FolderArea.Children.Remove(obj);
+                    Directory.Delete(@"UserFolder\" + obj.Text);
+                    string[] FoldersList = File.ReadAllLines(@"Folders.txt");
+                    File.Delete(@"Folders.txt");
+                    foreach (string folder in FoldersList)
                     {
-                        File.AppendAllText(@"Folders.txt", folder + "\n");
+                        string[] folderData = folder.Split('|');
+                        if (folderData[0] != obj.Text)
+                        {
+                            File.AppendAllText(@"Folders.txt", folder + "\n");
+                        }
                     }
                 }
-            }
-            else if (_currentUser._level >= Convert.ToInt32(obj.Tag))
-            {
-                FolderArea.Children.Remove(obj);
-                Directory.Delete(@"UserFolder\" + obj.Text);
-                string[] FoldersList = File.ReadAllLines(@"Folders.txt");
-                File.Delete(@"Folders.txt");
-                foreach (string folder in FoldersList)
+                else if (_currentUser._level >= Convert.ToInt32(obj.Tag))
                 {
-                    string[] folderData = folder.Split('|');
-                    if (folderData[0] != obj.Text)
+                    FolderArea.Children.Remove(obj);
+                    Directory.Delete(@"UserFolder\" + obj.Text);
+                    string[] FoldersList = File.ReadAllLines(@"Folders.txt");
+                    File.Delete(@"Folders.txt");
+                    foreach (string folder in FoldersList)
                     {
-                        File.AppendAllText(@"Folders.txt", folder + "\n");
-                    } else
-                    {
-                        File.Create(@"Folders.txt");
+                        string[] folderData = folder.Split('|');
+                        if (folderData[0] != obj.Text)
+                        {
+                            File.AppendAllText(@"Folders.txt", folder + "\n");
+                        }
+                        else
+                        {
+                            File.Create(@"Folders.txt");
+                        }
                     }
-                }
-            }
-            else
-            {
+                } else 
+                     {
                 MessageBox.Show("Вы не можете удалить эту папку", "Недостаточно прав");
-            }
+                     }
         }
 
         string folderName = "";
         private void EditFolder(object sender, RoutedEventArgs e)
         {
             var obj = ((sender as MenuItem).Parent as ContextMenu).PlacementTarget as DisplayTile;
-            if ((_currentUser._level >= Convert.ToInt32(obj.Tag)) || _currentUser._username == "Admin")
-            {
-                folderName = (((sender as MenuItem).Parent as ContextMenu).PlacementTarget as DisplayTile).Text;
-                Point DropPos = (((sender as MenuItem).Parent as ContextMenu).PlacementTarget as DisplayTile).PointToScreen(new Point(0, 0));
-                EditGrid.Visibility = Visibility.Visible;
-                Canvas.SetTop(EditGrid, DropPos.Y);
-                Canvas.SetLeft(EditGrid, DropPos.X);
-            }
-            else if (_currentUser._level >= Convert.ToInt32(obj.Tag))
-            {
-                folderName = (((sender as MenuItem).Parent as ContextMenu).PlacementTarget as DisplayTile).Text;
-                Point DropPos = (((sender as MenuItem).Parent as ContextMenu).PlacementTarget as DisplayTile).PointToScreen(new Point(0, 0));
-                EditGrid.Visibility = Visibility.Visible;
-                Canvas.SetTop(EditGrid, DropPos.Y);
-                Canvas.SetLeft(EditGrid, DropPos.X);
-            }
-            else
+                if ((_currentUser._level >= Convert.ToInt32(obj.Tag)) || _currentUser._username == "Admin")
+                {
+                    folderName = (((sender as MenuItem).Parent as ContextMenu).PlacementTarget as DisplayTile).Text;
+                    Point DropPos = (((sender as MenuItem).Parent as ContextMenu).PlacementTarget as DisplayTile).PointToScreen(new Point(0, 0));
+                    EditGrid.Visibility = Visibility.Visible;
+                    Canvas.SetTop(EditGrid, DropPos.Y);
+                    Canvas.SetLeft(EditGrid, DropPos.X);
+                }
+                else if (_currentUser._level >= Convert.ToInt32(obj.Tag))
+                {
+                    folderName = (((sender as MenuItem).Parent as ContextMenu).PlacementTarget as DisplayTile).Text;
+                    Point DropPos = (((sender as MenuItem).Parent as ContextMenu).PlacementTarget as DisplayTile).PointToScreen(new Point(0, 0));
+                    EditGrid.Visibility = Visibility.Visible;
+                    Canvas.SetTop(EditGrid, DropPos.Y);
+                    Canvas.SetLeft(EditGrid, DropPos.X);
+                }
+            else 
             {
                 MessageBox.Show("Вы не можете редактировать данную папку", "Недостаточно прав");
             }
@@ -193,9 +209,11 @@ namespace WindowsEmulator
                         {
                             string tag = (child as DisplayTile).Tag.ToString();
                             Directory.Delete(@"UserFolder\" + (child as DisplayTile).Text);
-                            string[] Folders = File.ReadAllLines(@"Folders.txt");
+
+                            string[] folders = File.ReadAllLines(@"Folders.txt");
                             File.Delete(@"Folders.txt");
-                            foreach (string folder in Folders)
+
+                            foreach (string folder in folders)
                             {
                                 string[] folderData = folder.Split('|');
                                 if ((child as DisplayTile).Text != folderData[0])
@@ -203,11 +221,14 @@ namespace WindowsEmulator
                                     File.AppendAllText(@"Folders.txt", folder + "\n");
                                 }
                             }
-                                (child as DisplayTile).Text = EditFolderName.Text;
+
+                            (child as DisplayTile).Text = EditFolderName.Text;
                             (child as DisplayTile).Tag = EditLevel.SelectedIndex + 1;
+
                             Point pos = (child as DisplayTile).PointToScreen(new Point(0, 0));
                             Directory.CreateDirectory(@"UserFolder\" + EditFolderName.Text);
-                            File.AppendAllText(@"Folders.txt", (child as DisplayTile).Text + "|" + (child as DisplayTile).Name + "|" + (child as DisplayTile).Tag.ToString() + "|" + (pos.X - 45) + "|" + (pos.Y - 60) + "\n");
+                            File.AppendAllText(@"Folders.txt", (child as DisplayTile).Text + "|" + (child as DisplayTile).Tag.ToString() + "|" + (pos.X - 45) + "|" + (pos.Y - 60) + "|_" + (child as DisplayTile).Name + "\n");
+
                             EditGrid.Visibility = Visibility.Hidden;
                             EditFolderName.Text = "";
                             EditLevel.SelectedIndex = 0;
@@ -267,14 +288,16 @@ namespace WindowsEmulator
                     newFolder.Text = FolderName.Text;
                     newFolder.Name = _currentUser._username;
                     newFolder.Tag = Level.SelectedIndex + 1;
+
                     FolderArea.Children.Add(newFolder);
                     Canvas.SetTop(newFolder, DropPos.Y - 60);
                     Canvas.SetLeft(newFolder, DropPos.X - 45);
+
                     if (!Directory.Exists(@"UserFolder\" + FolderName.Text))
                     {
                         Directory.CreateDirectory(@"UserFolder\" + FolderName.Text);
-                        File.AppendAllText(@"Folders.txt", FolderName.Text + "|" + newFolder.Name + "|" + newFolder.Tag.ToString() + "|"
-                            + (DropPos.X - 45).ToString() + "|" + (DropPos.Y - 60).ToString() + "\n");
+                        File.AppendAllText(@"Folders.txt", FolderName.Text + "|"  + newFolder.Tag.ToString() + "|"
+                            + (DropPos.X - 45).ToString() + "|" + (DropPos.Y - 60).ToString() + "|_" + newFolder.Name + "\n");
                         CreationGrid.Visibility = Visibility.Hidden;
                         FolderName.Text = "";
                         Level.SelectedIndex = 0;
@@ -291,7 +314,7 @@ namespace WindowsEmulator
                 }
                 else
                 {
-                    MessageBoxResult res = MessageBox.Show("Дайте папке имя, либо отмените создание\n", "Ошибка");
+                    MessageBoxResult res = MessageBox.Show("Дайте папке имя, либо отмените создание", "Ошибка");
                     FolderArea.Children.Remove(newFolder);
                     CreationGrid.Visibility = Visibility.Hidden;
                     FolderName.Text = "";
@@ -311,14 +334,36 @@ namespace WindowsEmulator
             string name = ((DisplayTile)sender).Name;
             ProcessStartInfo info = new ProcessStartInfo();
             info.FileName = ".\\UserFolder\\" + open;
-            if ((_currentUser._level >= Convert.ToInt32(tag)) || _currentUser._username == "Admin")
+            string names = name;
+            string[] name1 = names.Split('_');
+            foreach (string d in name1)
             {
-                Process.Start(info);
+                if (((_currentUser._level >= Convert.ToInt32(tag)) || d == _currentUser._username) || _currentUser._username == "Admin")
+                {
+                    Process.Start(info);
+                }
+                else if (_currentUser._level >= Convert.ToInt32(tag) || d == _currentUser._username)
+                {
+                    Process.Start(info);
+                }
             }
-            else if (_currentUser._level >= Convert.ToInt32(tag))
+        }
+        private void OpenDiscretionPanel(object sender, RoutedEventArgs e)
+        {
+            if (_currentUser._level >= 3)
             {
-                Process.Start(info);
-            } 
+                var task = new DiscretionPanel(_currentUser, this);
+                if (task.IsActive == false)
+                {
+                    task.ShowDialog();
+                    File.AppendAllText(@"Journal.txt", DateTime.Now.ToString("g") + " Вход в дискреционную панель: " + _currentUser._username + "\n");
+                }
+            }
+            else
+            {
+                MessageBox.Show("У вас нет доступа к дискреционной панели.", "Ошибка", MessageBoxButton.OK);
+                File.AppendAllText(@"Journal.txt", DateTime.Now.ToString(@"g") + " Попытка доступа к дискреционной панели: " + _currentUser._username + "\n");
+            }
         }
 
         private void OpenAdminPanel(object sender, RoutedEventArgs e)
@@ -375,7 +420,7 @@ namespace WindowsEmulator
             Canvas.SetLeft(obj, DropPos.X - 45);
             Canvas.SetTop(obj, DropPos.Y - 60);
             string[] FolderList = File.ReadAllLines(@"Folders.txt");
-            if (FolderList.Length != 0 && FolderList[0][0] != ' ' && obj.Text != "Журнал безопасности" && obj.Text != "Панель администратора")
+            if (FolderList.Length != 0 && FolderList[0][0] != ' ' && obj.Text != "Журнал безопасности" && obj.Text != "Панель администратора" && obj.Text != "Дискреционная панель")
                 {
                 File.Delete(@"Folders.txt");
                 foreach (string folder in FolderList)
@@ -386,7 +431,7 @@ namespace WindowsEmulator
                         File.AppendAllText(@"Folders.txt", FolderData[0] + "|" + FolderData[1] + "|" + FolderData[2] + "|" + FolderData[3] + "|" + FolderData[4] + "\n");
                     } else
                     {
-                        File.AppendAllText(@"Folders.txt", FolderData[0] + "|" + FolderData[1] + "|" + FolderData[2] + "|" + (DropPos.X - 45).ToString() + "|" + (DropPos.Y - 60).ToString() + "\n");
+                        File.AppendAllText(@"Folders.txt", FolderData[0] + "|" + FolderData[1] + "|" + (DropPos.X - 45).ToString() + "|" + (DropPos.Y - 60).ToString() + "|" + FolderData[4] + "\n");
                     }
                     }
                 }
